First version of an Integer only Rocket editor.
Treat track values as 32-bit ints.
Display values in hex or decimal.
Display values as 1x 32-bit, 2x 16-bit or 4x 8-bit numbers. (Optionally 8x 4-bit values but this takes a lot of room...)
Edit in hex values.
Support for new integer interpolation modes - LINEAR, SHORT, CHAR, NIBBLE.
Disable bias and scale shortcuts for now.

+++ src/Commands.c
@@ -534,7 +534,7 @@ static void toggleMute(void* userData) {
         struct track_key defKey;
         int i;
         size_t keysSize = sizeof(struct track_key) * data->syncTrack->num_keys;
-        float currentValue = (float)sync_get_val(data->syncTrack, data->row);
+        int currentValue = sync_get_val(data->syncTrack, data->row);
 
         data->track->disabled = true;
